Move tests inside ohsmart app
