+++ apps/ohsmart/package.json
@@ -10,12 +10,12 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@dans-framework/user-auth": "workspace:*",
     "@dans-framework/deposit": "workspace:*",
-    "@dans-framework/utils": "workspace:*",
     "@dans-framework/layout": "workspace:*",
     "@dans-framework/pages": "workspace:*",
     "@dans-framework/theme": "workspace:*",
+    "@dans-framework/user-auth": "workspace:*",
+    "@dans-framework/utils": "workspace:*",
     "@fontsource/roboto": "^5.0.7",
     "@mui/material": "^5.14.3",
     "i18next": "^23.4.1",
@@ -28,6 +28,8 @@
     "web-vitals": "^3.4.0"
   },
   "devDependencies": {
+    "@playwright/test": "^1.42.1",
+    "@types/node": "^20.11.27",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@vitejs/plugin-react": "^4.0.3",
