+++ apps/ohsmart/tests/mock-data/rights.json
