+++ apps/ohsmart/tests/pages/rights.page.ts
