+++ package.json
@@ -21,7 +21,6 @@
   "devDependencies": {
     "@dans-framework/eslint-config": "workspace:*",
     "@dans-framework/typescript-config": "workspace:*",
-    "@playwright/test": "^1.41.2",
     "@types/node": "^18.0.6",
     "eslint": "^8.54.0",
     "prettier": "^3.1.0",
